ext4: revert Huawei changes to ialloc
This is causing disasters in 4.3 and especially while SELinux
is enabled. System files are getting deleted and data is getting
remounted as read only. It seems to be the same files getting
deleted every time, common examples are:
- /system/app/Phonesky.apk (!)
- /system/media/bootanimation.zip (uses AOSP animation instead)
- /system/media/video/
- /system/wifi/ (!)

This is also present in the 3.0.8 kernel, but I'm not really
sure why it was changed in the first place - system remounting
seems to be working fine for me...

Not sure if recoveries will need to be rebuilt with this.